Progress for ensuring the app that might handle a URL is always available
- ContentView keeps a list of ResolveInfo items that can display a URL
- Settings now implements getAppsThatHandleUrl(). Soon this will check
against the user’s preferred apps for a given domain
- Move loadResolveInfoIntent() into MainApplication

This is a problem in the case like so (when user has ‘automatcally
load’ disabled).
1). G+ link loaded.
2). G+ link redirects to plus.google.com/appbasic/blah
3). “/appbasic/blah” is not handled by the G+ app, but the initial URL
is, so we need to be able to jump to the app. This is also a problem
with Twitter and YouTube links when the link is redirected to a mobile
link not handled by the app.