Incorrect ContentView often displays when coming from BubbleView, fixes #180, re #170
- MainController.onOpenUrl() takes setAsCurrentBubble. Prevents
multiple items being set as the current Bubble during a restore
- Add getBubbleIndex() to MainController and getIndexOfView() to
BubbleFlowView
- Remove duplicate calls to startScrollFinishedCheckTask() from
BubbleFlowView
- startScrollFinishedCheckTask() takes the targetX position as an
argument. Sometimes* getScrollX() is different to the desired targetX,
so this caters for that.

* Repro case: Load 5 bubbles via restore with ‘Auto content display’
set to ‘Always’, have a quick to load url last (e.g., techmeme.com/m/),
and observe what happens when that goes to display