Make restore previous session prompt show for longer
Chris mentioned some people complained about this in support. The
restore previous session comes up and goes away too fast. I think
probably because there is too much going on at the time the prompt
shows, there's not enough time to take that in + read the message.

Changing this to a long delay (6s) to address that.