Fix closeTab crash
This is the top crash in fabric here:
https://fabric.io/brave6/android/apps/com.linkbubble.playstore/issues/55c93db6e0d514e5d6e4b8ce

For wahtever reason the view is sometimes null when closing a tab. This
just checks to make sure it's not null.  Some of the checks that call
into closeTab already did null checks, the one that was crashing did
not. I thought it was better to just have the check inside the function.