Reduce allocations during bubble moving
If a bubble was dragged for a long time before it would store and keep
references to tha tnumber of events.  It would also do a lot of
allocations while the bubble was being dragged.  This reduces the
changes that garbage collection will be done at critical times.

We'll definitely need more optimizations and this does nothing to
address #520 which I think is a big source of people's negative
perceptions.