Attempt to prevent 301 redirects from loading.
This allows users to navigate backwards through 301 redirects by preventhing them from loading. Currently 301s load when navigating back, which makes it difficult to close bubbles from twitter. Unfortunately this has the side-effect of showing a white page, when instead we should probably keep navigating back. We probably don't want to land this either without a notice, and potentially forward button as the user can still resture the bubble in a blank state.

This is an attempt at fixing #562.