Attempt to fix back button traversing over 301 redirects.
This restores the custom URL stack keeping that we were previously doing, but augments how items are added to the list. This is generally a fairly dependable way of keeping track of the right URLs that we want, but currently it only adds a URL once the page is fully loaded, this could result in us missing a page in history if the user navigates away after it's done loading.