Add tracking protection option
This uses the disconnect list to enable tracking protection.

Full on ad block is still pending me finishing the C++ ad block plus
filter parser library and will land independent of this.

The plan is to have both options independent and both defaulted to off
until we rebrand.  We'll probably release both this and the adblock
together.