Automatically load LinkBubble if it's found as a valid target. Refs #579.
This might be a quick fix for #579, though it doesn't feel like it's the proper one. The problem is that Settings:getAppsThatHandleUrl is now returning additional apps, and I haven't tracked down exactly why that is the case. I would like to stop the bleeding asap, and I think this might be one way of doing it. I'm not sure if the default apps are functioning properly with this approach.