Format javascript according to semi-standard spec.
Using semi-standard for now due to the way that the JS is stringified and exectued within WebViews. In this case it breaks without the semi-colons. We'll fix this when time permits, but this works for now and is not worth spending more time on. At least the majority of the syntax matches other projects.