Optimized battery usage for expanded and at rest bubbles. Fixes #575
Updates would constantly be scheduled even for at rest and expanded
bubbles because mTouchView can be non null sometimes.   The cehck for
scheduling an update is now also be gated on mStillTouchFrameCount > -1.

The problem was introduced in January 2014

Was introduced for a fix to this issue:
https://github.com/brave/link-bubble/issues/189

Changeset which introduced it:
https://github.com/brave/link-bubble/commit/1a831eb45e487162c0748e59864f49a01b612d7a