Optimize / simplify dragging code for less lag
- Using the absolute positions in onActionMove for move events to avoid
  out of sync finger and bubble. Fixes #520.
- Simplify / remove some animations which were causing laggy bubble
 dragging. Fixes #488
- updateViewLayout which is called from
 MainController.updateRootWindowLayout is slow and calls should be
 minimized.  We never call it with the same params twice anymore,
 and also we reduce the amount of times it’s called when dragging.
- Simplified target bubbles (close, share left/right) handling so
 that animations don’t happen when dragging.  In particular it
 would set a target position constantly based on the bubble
 position and then it would call update on each frame which would
 call the slow updateViewLayout.

The only difference I can see visually is that the target bubbles
(close, share left, share right) would move around before and they don't
anymore.  For me this seems to run a lot more smoohtly. The main
difference is we try to avoid calls to updateViewLayout while dragging
is happening.

When Constants.PROFILE_FPS is true the number of bad frames (slower than
40FPS) used to grow increasingly forewver, but it doesn't any longer.