Various tweaks to use automatic animations
This changeset fixes these problems:
- When you have a bubble on the side, tapping it should just fade in the
webview, currently it translates the webview up.
  https://app.asana.com/0/64274785767353/46881973101038
- When you tap a bubble on top it should fade out the webview and move
  the bubble back to its original tap position.
  https://app.asana.com/0/64274785767353/64274785767360
- If you start dragging but then stop, it should fade page back in.
  https://app.asana.com/0/64274785767353/64276992253112

In particular the changes do less animations manually in software on
each doFrame and instead submits the animations to Android to do.  The
manual ones can't effectively be hardware accelerated.

The way in general to have performant bubbles is to continue this trend
and do do as little as possible in the doFrame draw loop.

Independent of those tasks, I spent a long time also trying to move over
translations of bubbles to use those animations too but didn't have any
luck.  The problem is that you can't translate a root view with
animations without having it clip to the current view's size.  The main
draggable bubble is a root view.  It does work and is lightning fast if
you put the view into a ViewGroup first like FrameLayout or
RelativeLayout.

If I put it in a ViewGroup I get super fast operations for everything,
but then there is no way to have clicks go to background applications.
You can set it as an overlay, but then every child element is also
treated the same and can't get touch events.  I also tried dynamically
changing the ViewGroup or dimensions of the view only when dragging but neither worked.